GenBin only prompts if bin exists

+++ firmware/genBin.sh
@@ -19,10 +19,13 @@ export PICO_SDK_PATH=$SCRIPT_DIR/lib/pico-sdk
 # Tell CMake what we're building for
 export PICO_BOARD_HEADER_DIRS=$SCRIPT_DIR
 
-# Clear bin folder and create a new one
-echo "Removing $SCRIPT_DIR/bin..."
-rm -rI bin # I dont trust my bash skills, so you're getting prompted.
-mkdir bin  # Cope.
+if [ -d bin ]
+   then
+       # Clear bin folder and create a new one
+       echo "Removing $SCRIPT_DIR/bin..."
+       rm -rI bin # I dont trust my bash skills, so you're getting prompted.
+fi
+mkdir bin
 
 # Make the makefiles
 cd bin
